Correction for accessing the members of stomp_session_t from third-party applications
