+++ src/frame.c
@@ -20,49 +20,8 @@
 #include <errno.h>
 #include <unistd.h>
 
-#include "frame.h"
-#include "hdr.h"
-
-enum read_state {
-	RS_INIT,
-	RS_CMD,
-	RS_HDR,
-	RS_HDR_ESC,
-	RS_BODY,
-	RS_DONE,
-	RS_ERR
-};
-
-struct frame_hdr {
-	ptrdiff_t key_offset;
-	size_t key_len;
-	ptrdiff_t val_offset;
-	size_t val_len;
-};
-
-struct _frame {
-	void *buf;
-	size_t buf_len; /* length of data in buf in bytes */
-	size_t buf_capacity; /* allocated size in bytes */
-
-	ptrdiff_t cmd_offset; /* offset in buff to the start of the cmd string */
-	size_t cmd_len; /* lenght of cmd string in bytes */
-
-	struct frame_hdr *hdrs; /* array of struct frame_hdr elements */
-	size_t hdrs_len; /* number of elements in the array */
-	size_t hdrs_capacity; /* allocated number of struct frame_hdr elements */
-
-	struct stomp_hdr *stomp_hdrs; /* array of struct stomp_hdr elements */
-	size_t stomp_hdrs_len; /* number of elements in the array */
-	size_t stomp_hdrs_capacity; /* allocated number of struct stomp_hdr elements */
-
-	ptrdiff_t body_offset; /* offset in buff to the start of the body */
-	size_t body_len; /* length of body in bytes */
-
-	enum read_state read_state; /* current state of the frame reading state mashine */
-	ptrdiff_t tmp_offset; /* current position within buf while reading an incomming frame */
-	size_t tmp_len; /* amount of bytes read while reading an incomming frame */
-};
+#include <stomp/frame.h>
+#include <stomp/hdr.h>
 
 /* number of bytes to increase session->buf by
  * when adding more data to the frame */
